### 2019-07-05 #19 @whedon
[PRE REVIEW]: Econometrics.jl
**Submitting author:** @Nosferican (<a href="http://orcid.org/0000-0002-8406-6175">José Bayoán Santiago Calderón</a>)
**Repository:** <a href="https://github.com/Nosferican/Econometrics.jl" target ="_blank">https://github.com/Nosferican/Econometrics.jl</a>

Thanks for submitting your paper to JuliaCon @Nosferican. The JuliaCon editor @matbesancon, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@Nosferican if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
