### 2019-07-15 #38 @whedon
[REVIEW]: Econometrics.jl
**Submitting author:** @Nosferican (<a href="http://orcid.org/0000-0002-8406-6175">José Bayoán Santiago Calderón</a>)
**Repository:** <a href="https://github.com/Nosferican/Econometrics.jl" target ="_blank">https://github.com/Nosferican/Econometrics.jl</a>
