### 2019-07-15 f49a22d
Added JuliaCon shield.
