### 2019-07-22 #38 @cpfiffer
[econometrics-jl-comments.pdf](https://github.com/JuliaCon/proceedings-review/files/3419300/econometrics-jl-comments.pdf)

I've uploaded a PDF with some of my comments on the proof. Mostly minor stuff.

The remaining points I have on the review list are

- API documentation. Your docstrings are pretty light in the actual source, and you don't have Documenter.jl set up to publish the docstrings on the site. I'd put a little time in to that if you have any.
- Community guidelines. The review checklist asks for this and it's not in the repo, so I'm not going to check it off. I'm also not super sure how important it is.

Generally good work. I had a fairly pleasant time working with it -- most everything seems quick and easy to use, and it's easy to shift between standard error types.
