### 2019-07-23 6b6cc0e
Create CODE_OF_CONDUCT.md
