### 2019-07-23 6b2ec9c
Docstring every function
