+++ src/structs.jl
@@ -1,12 +1,38 @@
+"""
+    EconometricsModel <: RegressionModel
+
+Abstract type provided by Econometrics.jl
+"""
 abstract type EconometricsModel <: RegressionModel end
 implicit_intercept(::Type{<:EconometricsModel}) = true
 # Estimators
+"""
+    ModelEstimator
+
+Abstract type for model estimators.
+"""
 abstract type ModelEstimator end
+"""
+    LinearModelEstimators
+
+Abstract type for linear model estimators.
+"""
 abstract type LinearModelEstimators <: ModelEstimator end
+"""
+    ContinuousResponse(groups::Vector{Vector{Vector{Int}}}) <: LinearModelEstimators
+
+Continuous response estimator with potential features absorption.
+"""
 struct ContinuousResponse <: LinearModelEstimators
     groups::Vector{Vector{Vector{Int}}}
 end
 show(io::IO, estimator::ContinuousResponse) = println(io, "Continuous Response Model")
+"""
+    BetweenEstimator(effect::Symbol,
+                     groups::Vector{Vector{Int}}) <: LinearModelEstimators
+
+Continuous response estimator collapsing a dimension in a longitudinal setting.
+"""
 struct BetweenEstimator <: LinearModelEstimators
     effect::Symbol
     groups::Vector{Vector{Int}}
@@ -26,6 +52,15 @@ function show(io::IO, estimator::BetweenEstimator)
         end
     end
 end
+"""
+    RandomEffectsEstimator(pid::Tuple{Symbol,Vector{Vector{Int}}},
+                           tid::Tuple{Symbol,Vector{Vector{Int}}},
+                           idiosyncratic::Float64,
+                           individual::Float64,
+                           θ::Vector{Float64}) <: LinearModelEstimators
+
+Swamy-Arora estimator.
+"""
 struct RandomEffectsEstimator <: LinearModelEstimators
     pid::Tuple{Symbol,Vector{Vector{Int}}}
     tid::Tuple{Symbol,Vector{Vector{Int}}}
@@ -72,6 +107,11 @@ function show(io::IO, estimator::RandomEffectsEstimator)
     println(io, "idiosyncratic error component: $(round(idiosyncratic, digits = 4))")
     println(io, "ρ: $(round(individual^2 / (individual^2 + idiosyncratic^2), digits = 4))")
 end
+"""
+    NominalResponse(categories::ContrastsMatrix) <: ModelEstimator
+
+Multinomial logistic regression.
+"""
 struct NominalResponse{T} <: ModelEstimator
     categories::Vector{T}
     function NominalResponse(obj::ContrastsMatrix)
@@ -85,6 +125,11 @@ function show(io::IO, estimator::NominalResponse)
     println(io, "Probability Model for Nominal Response")
     println(io, "Categories: $(join(estimator.categories, ", "))")
 end
+"""
+    OrdinalResponse(categories::ContrastsMatrix) <: ModelEstimator
+
+Proportional odds logistic regression.
+"""
 struct OrdinalResponse{T} <: ModelEstimator
     categories::Vector{T}
     function OrdinalResponse(obj::ContrastsMatrix)
@@ -98,7 +143,11 @@ function show(io::IO, estimator::OrdinalResponse)
     println(io, "Probability Model for Ordinal Response")
     println(io, "Categories: $(join(estimator.categories, " < "))")
 end
-# Variance-covariance Estimators
+"""
+    VCE
+
+Variance-covariance estimators.
+"""
 @enum VCE begin
     OIM = -1
     HC0 = 0
