### 2019-07-23 e202461
Add compat for latest Documenter
