+++ docs/Project.toml
@@ -3,4 +3,4 @@ CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 RDatasets = "ce6b1742-4840-55fa-b093-852dadbb1d8b"
 [compat]
-Documenter = "0.22"
+Documenter = "0.22, 0.23"
