### 2019-07-23 #38 @Nosferican
Thanks for taking time to go over it. I took a brief look at the PDF with the comments and will be making revisions in the next couple days.

I have updated the repository to docstring every functioned and struct defined in the package. Most of the functions lack docstring if those are defined in other packages and their docstring are appropriate.

The public API includes: `EconometricModel`, `absorb`, `BetweenEstimator`, `RandomEffectsEstimator`, `OIM`, `HC0`, `HC1`, `HC2`, `HC3`, `HC4`. In addition to the public API, there are a few `StatsBase.RegressionModel` API methods with modified methods (e.g., `fit` and `vcov`). Let me know if any are missing from the documentation.

I have also added the community guidelines to the repo.
