### 2019-07-23 #38 @cpfiffer
For the documentation API I was referring to using Documenter's abiity to export docstrings, not whether or not you have a documentation site. Currently it contains some useful examples examples, but no API library.

Using the [StatsModels](https://juliastats.github.io/StatsModels.jl/stable/api/) site as an example, they have an `api.md` that uses Documenter to render all their docstrings to the website. It's pretty straightforward to add it to your site -- here's the [StatsModels api.md file](https://github.com/JuliaStats/StatsModels.jl/blob/master/docs/src/api.md) and their [make file](https://github.com/JuliaStats/StatsModels.jl/blob/master/docs/make.jl).

Doing this would make it a lot easier for people to click around your documentation site and see how certain things work, what they need, what they do, etc.

At this point, it's only really a suggestion. I've checked off the last box for the review. I think having the docstrings as they are right now is satisfactory, but I would recommend publishing the API on the site going forward.

Great work!
