### 2019-07-23 #38 @Nosferican
Got it. Thanks for the feedback. I think the public API is stable enough to make it "official" and publish it in the docs now (e.g., the internals of the types). Will do.
