### 2019-07-29 ff5dbd4
Fix pretty print
Handles the string representation of formula based on the user input and model type.
