### 2019-08-03 #38 @pkofod
I have a few comments for the paper and repo. Note that I did not check off all boxes as I do not believe everything is totally fulfilled.

**Installation**
There are two boxes that relate to installation: one for the docs and one for the paper. Neither provide instructions about how to actually install the software.

**Statement of need**
The paper does clearly state what the point of the software is, but the documentation actually doesn't.

**DOIs**
Not all papers, books, etc have DOIs. One example is https://www.springer.com/kr/book/9783540755708 that has no doi listed, but actually has one.

**Page limit**
Is there a page limit? I couldn't find one, so I'm not sure. @matbesancon

**Paper**
Mostly good. I came across one paragraph that stood out as odd/unnecessary:
```
The Julia language [3] is an upcoming language especially well-
suited for scientific computing such as econometrics, data science,
machine learning, and other related tasks.
```
I suggest you delete it.

Could you briefly discuss/mention what the point of `EconometricModel` is? When the reader gets to the figures, these come a bit out of the blue; I tried to go back and see if I had missed their introduction! It's "shown" how to use the software, but not really explained. I'm not requesting 2 pages, but just something that says what goes where, etc. It's not clear to the uninitiated - though I obviously don't expect a full "documentation" repeated in the paper.
