### 2019-08-04 2ba7b13
Implements the Tables.jl API
- Drops DataFrames.jl
- Adheres to CategoricalArrays
