### 2019-08-04 09786e5
Implements the Tables.jl API (#13)
- Drops DataFrames.jl
- Adheres to CategoricalArrays
