### 2019-08-04 #38 @Nosferican
Thanks for the review.

**Paper Limit**
The normal paper is 6-8 pages according to https://github.com/JuliaCon/proceedings.juliacon.org/pull/9#issuecomment-511241172.

**On the issue of DOI for references.**
The DOI are included in the source (see [ref.bib](https://github.com/Nosferican/Econometrics.jl/blob/master/paper/ref.bib#L393)), but the JuliaCon template does not display those (see https://github.com/JuliaCon/JuliaConSubmission.jl/issues/7).

I will be working on incorporating the feedback I received from y'all next week. I will ping y'all and share the revisions. Thanks!
