### 2019-08-04 a001dba
Refractor of struct
A rough refactor such that models adhere to the `fit!`/`fit` paradigm.
