+++ src/solvers.jl
@@ -15,6 +15,50 @@ For example, Swamy-Arora random effects model for longitudinal data.
                       z::AbstractVecOrMat{<:Number},
                       Z::AbstractMatrix{<:Number},
                       wts::AbstractVector)
+	if !isa(wts, Weights)
+        wts = FrequencyWeights(wts)
+    else
+        isa(wts, FrequencyWeights) || throw(ArgumentError("Only frequency weights are supported."))
+    end
+    X = transform(estimator, X, wts)
+    y = transform(estimator, y, wts)
+    z = transform(estimator, z, wts)
+    Z = transform(estimator, Z, wts)
+    wts = transform(estimator, FrequencyWeights(wts))
+    if !isempty(z)
+        Z̃ = hcat(X, Z)
+        F = bunchkaufman(Hermitian(Z̃' * Diagonal(wts) * Z̃), true, check = false)
+        bkr = count(x -> abs(x) ≥ √eps(), F.D)
+		if bkr < size(F, 2)
+            lin_ind = sort!(invperm(F.p)[1:bkr])
+            count(x -> lin_ind > size(X, 2) + 1) ≥ size(z, 2) ||
+                throw(ArgumentError("Insufficient number of instruments."))
+            Z̃ = Z̃[:,lin_ind]
+            F = bunchkaufman(Hermitian(Z̃' * Diagonal(wts) * Z̃), true)
+        else
+            lin_ind = collect(1:size(F, 2))
+        end
+        γ = F \ (Z̃' * Diagonal(wts) * z)
+        X̃ = hcat(X, Z̃ * γ)
+    else
+        X̃ = X
+    end
+    F = bunchkaufman(Hermitian(X̃' * Diagonal(wts) * X̃), true, check = false)
+    bkr = count(x -> abs(x) ≥ √eps(), F.D)
+    if bkr < size(F, 2)
+        lin_ind = sort!(invperm(F.p)[1:bkr])
+        X̃ = convert(Matrix{Float64}, X̃[:,lin_ind])
+        F = bunchkaufman(Hermitian(X̃' * Diagonal(wts) * X̃), true)
+    else
+        lin_ind = collect(1:size(F, 2))
+    end
+    β = F \ (X̃' * Diagonal(wts) * y)
+    Ψ = Hermitian(inv(F))
+    ŷ = isempty(z) ? X̃ * β : hcat(X, z)[:,lin_ind] * β
+	X̃, y, β, Ψ, ŷ, wts, lin_ind
+end
+@views function fit!(obj::EconometricModel{<:LinearModelEstimators})
+	@unpack estimator, X, y, z, Z, wts = obj
     if !isa(wts, Weights)
         wts = FrequencyWeights(wts)
     else
@@ -24,10 +68,10 @@ For example, Swamy-Arora random effects model for longitudinal data.
     y = transform(estimator, y, wts)
     z = transform(estimator, z, wts)
     Z = transform(estimator, Z, wts)
-    w = transform(estimator, FrequencyWeights(wts))
+    wts = transform(estimator, FrequencyWeights(wts))
     if !isempty(z)
         Z̃ = hcat(X, Z)
-        F = bunchkaufman(Hermitian(Z̃' * Diagonal(w) * Z̃), true, check = false)
+        F = bunchkaufman(Hermitian(Z̃' * Diagonal(wts) * Z̃), true, check = false)
         bkr = count(x -> abs(x) ≥ √eps(), F.D)
 		if bkr < size(F, 2)
             lin_ind = sort!(invperm(F.p)[1:bkr])
@@ -38,41 +82,31 @@ For example, Swamy-Arora random effects model for longitudinal data.
         else
[…]
