### 2019-08-04 2de3aad
Refractor of struct
A rough refactor such that models adhere to the `fit!`/`fit` paradigm.
