### 2019-08-04 af3fcbd
Refractor of struct (#14)
A rough refactor such that models adhere to the `fit!`/`fit` paradigm.
