+++ src/main.jl
@@ -27,15 +27,48 @@ mutable struct EconometricModel{E<:ModelEstimator,
     data::T
     X::Matrix{Float64}
     y::Y
-    w::W
+    wts::W
     β::Vector{Float64}
     Ψ::Hermitian{Float64,Matrix{Float64}}
     ŷ::Ŷ
+    z::Matrix{Float64}
+    Z::Matrix{Float64}
     vars::N
     iv::Int
     vce::VC
+    function EconometricModel(estimator::Type{<:Union{EconometricModel,ModelEstimator}},
+                              f::FormulaTerm,
+                              data;
+                              contrasts::Dict{Symbol} = Dict{Symbol,Union{<:AbstractContrasts,<:AbstractTerm}}(),
+                              wts::Union{Nothing,Symbol} = nothing,
+                              panel::Union{Nothing,Symbol} = nothing,
+                              time::Union{Nothing,Symbol} = nothing,
+                              vce::VCE = OIM)
+        data, exogenous, iv, estimator, X, y, z, Z, wts =
+            decompose(deepcopy(f), data, contrasts, wts, panel, time, estimator, vce)
+        if isa(estimator, Union{NominalResponse, OrdinalResponse})
+            @unpack categories = estimator
+            y = [ findfirst(isequal(x), categories) for x ∈ y ]
+            @assert length(categories) > 2
+        end
+        if isa(estimator, NominalResponse)
+            ŷ = zeros(0, 0)
+        else
+            ŷ = zeros(0)
+        end
+        wts = FrequencyWeights(collect(wts))
+        vars = (coefnames(exogenous.lhs),
+                convert(Vector{String}, vcat(coefnames(exogenous.rhs), coefnames(iv.lhs))))
+        new{typeof(estimator), typeof(f), typeof(data), typeof(y), typeof(wts), typeof(ŷ), typeof(vars), typeof(vce)}(
+            estimator, f, data, X, y, wts, zeros(0), Hermitian(zeros(0, 0)), ŷ,
+            z, Z, vars, size(Z, 2), vce)
+    end
 end
 function show(io::IO, obj::EconometricModel{<:LinearModelEstimators})
+    if !isfitted(obj)
+        println(io, "Model has not been fitted.")
+        return obj
+    end
     show(io, obj.estimator)
     println(io, @sprintf("Number of observations: %i", nobs(obj)))
     println(io, @sprintf("Null Loglikelihood: %.2f", nullloglikelihood(obj)))
@@ -50,6 +83,10 @@ function show(io::IO, obj::EconometricModel{<:LinearModelEstimators})
     show(io, coeftable(obj))
 end
 function show(io::IO, obj::EconometricModel{<:ContinuousResponse})
+    if !isfitted(obj)
+        println(io, "Model has not been fitted.")
+        return obj
+    end
     show(io, obj.estimator)
     ℓℓ₀ = nullloglikelihood(obj)
     ℓℓ = loglikelihood(obj)
@@ -78,6 +115,10 @@ function show(io::IO, obj::EconometricModel{<:ContinuousResponse})
     show(io, coeftable(obj))
 end
 function show(io::IO, obj::EconometricModel{<:NominalResponse})
+    if !isfitted(obj)
+        println(io, "Model has not been fitted.")
+        return obj
+    end
     show(io, obj.estimator)
     ℓℓ₀ = nullloglikelihood(obj)
     ℓℓ = loglikelihood(obj)
@@ -97,6 +138,10 @@ function show(io::IO, obj::EconometricModel{<:NominalResponse})
     show(io, coeftable(obj))
 end
 function show(io::IO, obj::EconometricModel{<:OrdinalResponse})
+    if !isfitted(obj)
[…]
