### 2019-08-06 4020a73
Updating docs
- Incremental improvement to docs
- CI to include Win64/Win32
- Improve clean formula
- Make API confint consistent
