### 2019-08-06 6b17985
Update .travis.yml
- See if it uses the correct version now
