### 2019-08-06 9fb1cc6
Fix Project
Added note about CI OS/Architecture.
