### 2019-08-06 09d606d
Updating docs (#16)
Incorporate recommendation from peer-review at https://github.com/JuliaCon/proceedings-review/issues/38.

- Statement of need through the docs
- Installation Guide
- Document API
- Explain the role of the dispatcher on model type

- Update CI for Windows
- bug fix of pretty formula
- consistent use of `level` in `confint`
