+++ .travis.yml
@@ -1,15 +1,27 @@
 language: julia
-os:
-  - linux
-julia:
-  - 1.1
+matrix:
+  include:
+    - name: linux
+      os: linux
+      julia: 1.3
+    - name: osx
+      os: osx
+      julia: 1.3
+    - name: win64
+      os: windows
+      arch: x64
+      julia: 1.3
+    - name: win32
+      os: windows
+      arch: x86
+      julia: 1.3
 codecov: true
 notifications:
   email: false
 jobs:
   include:
     - stage: "Documentation"
-      julia: 1.1
+      julia: 1.3
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
