+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#Econometrics-1","page":"Econometrics","title":"Econometrics","text":"","category":"section"},{"location":"#Setup-1","page":"Econometrics","title":"Setup","text":"","category":"section"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"using Econometrics, CSV, RDatasets","category":"page"},{"location":"#Continuous-Response-Models-1","page":"Econometrics","title":"Continuous Response Models","text":"","category":"section"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"data = RDatasets.dataset(\"Ecdat\", \"Crime\") |>\n  (data -> select(data, [:County, :Year, :CRMRTE, :PrbConv, :AvgSen, :PrbPris]))\nfirst(data, 6)","category":"page"},{"location":"#Pooling-1","page":"Econometrics","title":"Pooling","text":"","category":"section"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"model = fit(EconometricModel,\n            @formula(CRMRTE ~ PrbConv + AvgSen + PrbPris),\n            data)","category":"page"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"You can also request heteroscedasticity consistent estimators for linear models.","category":"page"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"vcov(model, HC0)","category":"page"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"vcov(model, HC1)","category":"page"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"vcov(model, HC2)","category":"page"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"vcov(model, HC3)","category":"page"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"vcov(model, HC4)","category":"page"},{"location":"#Between-1","page":"Econometrics","title":"Between","text":"","category":"section"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"model = fit(BetweenEstimator,\n            @formula(CRMRTE ~ PrbConv + AvgSen + PrbPris),\n            data,\n            panel = :County)","category":"page"},{"location":"#Within-1","page":"Econometrics","title":"Within","text":"","category":"section"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"model = fit(EconometricModel,\n            @formula(CRMRTE ~ PrbConv + AvgSen + PrbPris + absorb(County)),\n            data)","category":"page"},{"location":"#Within-(two-ways)-1","page":"Econometrics","title":"Within (two-ways)","text":"","category":"section"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"model = fit(EconometricModel,\n            @formula(CRMRTE ~ PrbConv + AvgSen + PrbPris + absorb(County + Year)),\n            data)","category":"page"},{"location":"#Random-1","page":"Econometrics","title":"Random","text":"","category":"section"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"model = fit(RandomEffectsEstimator,\n            @formula(CRMRTE ~ PrbConv + AvgSen + PrbPris),\n            data,\n            panel = :County,\n            time = :Year)","category":"page"},{"location":"#Random-with-Instrumental-Variables-1","page":"Econometrics","title":"Random with Instrumental Variables","text":"","category":"section"},{"location":"#","page":"Econometrics","title":"Econometrics","text":"model = fit(RandomEffectsEstimator,\n            @formula(CRMRTE ~ PrbConv + (AvgSen ~ PrbPris)),\n            data,\n            panel = :County,\n            ti
[…]
