### 2019-08-06 5df14db
Update .travis.yml
Make sure that the jobs are properly named.
