+++ .travis.yml
@@ -4,18 +4,14 @@ matrix:
   include:
     - os: linux
       julia: 1.3
-      name: "linux"
     - os: osx
       julia: 1.3
-      name: "osx"
     - os: windows
       arch: x64
       julia: 1.3
-      name: "win64"
     - os: windows
       arch: x86
       julia: 1.3
-      name: "win32"
 codecov: true
 notifications:
   email: false
