### 2019-08-06 0fbd053
Update .travis.yml
Check if valid travis.yml
