+++ .travis.yml
@@ -1,24 +1,21 @@
 language: julia
-matrix:
-  fast_finish: true
-  include:
-    - os: linux
-      julia: 1.3
-    - os: osx
-      julia: 1.3
-    - os: windows
-      arch: x64
-      julia: 1.3
-    - os: windows
-      arch: x86
-      julia: 1.3
+julia: 1.3
 codecov: true
 notifications:
   email: false
 jobs:
   include:
+    - stage: "Test"
+      matrix:
+        fast_finish: true
+        include:
+          - os: linux
+          - os: osx
+          - os: windows
+            arch: x64
+          - os: windows
+            arch: x86
     - stage: "Documentation"
-      julia: 1.3
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
