+++ .travis.yml
@@ -1,22 +1,23 @@
 language: julia
-julia: 1.3
-codecov: true
+os:
+  - linux
+  - osx
+julia:
+  - 1.3
 notifications:
   email: false
 jobs:
   include:
-    - stage: "Test"
-      matrix:
-        fast_finish: true
-        include:
-          - os: linux
-          - os: osx
-          - os: windows
-            arch: x64
-          - os: windows
-            arch: x86
+    - os: windows
+      arch: x64
+      julia: 1.3
+    - os: windows
+      arch: x86
+      julia: 1.3
+      name: "Win 32-bit"
     - stage: "Documentation"
       os: linux
+      julia: 1.3
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
                                                Pkg.instantiate()'
