+++ .travis.yml
@@ -2,6 +2,7 @@ language: julia
 os:
   - linux
   - osx
+  - windows
 julia:
   - 1.3
 notifications:
@@ -9,12 +10,8 @@ notifications:
 jobs:
   include:
     - os: windows
-      arch: x64
-      julia: 1.3
-    - os: windows
       arch: x86
       julia: 1.3
-      name: "Win 32-bit"
     - stage: "Documentation"
       os: linux
       julia: 1.3
