### 2019-08-13 969ea35
Adds Windows and x86/x64
Travis now supports Windows. I include 32 and 64 bits architectures for Windows.
