+++ .travis.yml
@@ -1,28 +1,20 @@
 language: julia
-matrix:
-  include:
-    - name: linux
-      os: linux
-      julia: 1.3
-    - name: osx
-      os: osx
-      julia: 1.3
-    - name: win64
-      os: windows
-      arch: x64
-      julia: 1.3
-    - name: win32
-      os: windows
-      arch: x86
-      julia: 1.3
-codecov: true
+os:
+  - linux
+  - osx
+  - windows
+julia:
+  - 1.3
 notifications:
   email: false
 jobs:
   include:
-    - stage: "Documentation"
+    - os: windows
+      arch: x86
       julia: 1.3
+    - stage: "Documentation"
       os: linux
+      julia: 1.3
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
                                                Pkg.instantiate()'
