### 2019-08-13 0f1471f
Minor improvements to docs
