### 2019-08-13 954d81e
Delete index.md
Move it to `jmd`.
