+++ docs/src/index.md
@@ -1,3 +1,0 @@
-# Econometrics.jl
-
-Thanks for your interest in this project. Please use the menu for finding the relevant sections.
