+++ docs/make.jl
@@ -12,7 +12,8 @@ makedocs(format = Documenter.HTML(assets = ["assets/custom.css"]),
          sitename = "Econometrics.jl",
          pages = ["Introduction" => "index.md",
                   "Getting Started" => "getting_started.md",
-                  "Estimators" => "estimators.md"]
+                  "Estimators" => "estimators.md",
+                  "Public API" => "public_api.md"]
     )
 
 deploydocs(repo = "github.com/Nosferican/Econometrics.jl.git")
