### 2019-08-13 2eb04df
Minor improvements to docs
* Improvements to docs
* Public API
