+++ README.md
@@ -1,12 +1,17 @@
 # Econometrics.jl
 
-[![License: ISC - Permissive License](https://img.shields.io/badge/License-ISC-green.svg)](https://img.shields.io/badge/License-ISC-green.svg)
+[![License: ISC - Permissive License](https://img.shields.io/badge/License-ISC-green.svg)](https://img.shields.io/github/license/Nosferican/Econometrics.jl)
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![JuliaCon](https://submissions.juliacon.org/papers/446fde271579d85e0d4c691d54093dbb/status.svg)](https://submissions.juliacon.org/papers/446fde271579d85e0d4c691d54093dbb)
+
 [![Documentation: stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://nosferican.github.io/Econometrics.jl/stable)
 [![Documentation: dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://nosferican.github.io/Econometrics.jl/dev)
+
 [![Build Status](https://travis-ci.com/Nosferican/Econometrics.jl.svg?branch=master)](https://travis-ci.com/Nosferican/Econometrics.jl)
 [![Code Coverage](https://codecov.io/gh/Nosferican/Econometrics.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/Nosferican/Econometrics.jl)
-[![JuliaCon](https://submissions.juliacon.org/papers/446fde271579d85e0d4c691d54093dbb/status.svg)](https://submissions.juliacon.org/papers/446fde271579d85e0d4c691d54093dbb)
+
+![GitHub commits since latest release](https://img.shields.io/github/commits-since/Nosferican/Econometrics.jl/v0.1.0)
+
 
 This package uses continuous integration on Linux, OSX, and Windows (x64/x86).
 
