+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"getting_started/#Getting-Started-1","page":"Getting Started","title":"Getting Started","text":"","category":"section"},{"location":"getting_started/#Installation-1","page":"Getting Started","title":"Installation","text":"","category":"section"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"Econometrics.jl is registered with the Julia's General registry.","category":"page"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"For installing the package, users can use Julia's package manager.","category":"page"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"using Pkg; Pkg.add(\"Econometrics\") # one alternative\n]add Econometrics # same as the line above","category":"page"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"The package manager will automatically resolve which release version for Econometrics.jl is compatible with the environment in use.","category":"page"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"The package manager allows for other experiences such as downloading the development version of the package through","category":"page"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"]add Econometrics#master","category":"page"},{"location":"getting_started/#Data-1","page":"Getting Started","title":"Data","text":"","category":"section"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"Econometrics.jl relies on users being able to pass in data for estimating the various models. This package allows for data to be passed through any tabular representation that implements the Tables.jl API. For example, a DataFrame from DataFrames.jl is a common solution for in-memory tabular data that implements the Tables.jl API.","category":"page"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"The second assumption Econometrics.jl makes about the data is that categorical data uses the CategoricalArrays.jl representations (i.e., AbstractCategoricalVector, CategoricalValue).","category":"page"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"Lastly, the package currently assumes in-memory data.","category":"page"},{"location":"getting_started/#Specifying-a-Model-1","page":"Getting Started","title":"Specifying a Model","text":"","category":"section"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"Statistical models use a formula with formulae syntax to specify the relation among the various features. For example, Econometrics.jl uses the following syntax","category":"page"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"modelformula = @formula(response ~ exogenous + (endogenous ~ instruments) + absorb(high_dimensional_controls))","category":"page"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"where the right-hand side is the response and the left-hand side has","category":"page"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"exogenous variables\npotentially endogenous features and additional instruments\nhigh dimensional controls","category":"page"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"Formulas also allow for passing special term
[…]
