### 2019-08-13 3e87bae
Updates from Overleaf
