+++ paper/paper.tex
@@ -238,6 +238,8 @@ This section will showcase some examples of using the package for various estima
 
 For linear models, the examples use a crime dataset \cite{Crime}. The data set is a balanced longitudinal data set with 90 counties in North Carolina from 1981 to 1987. The outcome variable is the crime rate and the explanatory variables include the probability of conviction, average sentence, and probability of prison sentence. Estimating the pooling estimator or the between estimator can be accomplished as in figure \vref{estimator_pooling} and \vref{estimator_between}. Table \vref{poberes} shows the estimated 95\% confidence intervals using Econometrics.jl, Stata, and R's plm package.
 
+
+
 \begin{figure}[hbpt]
     \includegraphics[max width=\linewidth,center]{estimator_pooling}
 	\caption{Estimation of the pooling estimator}
