### 2019-08-13 449f8f0
- Added brief explanation of the `fit` method - Added a note about how to install the package
