### 2019-08-13 #38 @Nosferican
@cpfiffer & @pkofod I have updated the paper and docs to address the feedback. Thanks!
