### 2019-08-13 #38 @pkofod
@matbesancon can whedon re-compile the pdf ?
