### 2019-08-14 #38 @pkofod
Can you perhaps tell me in what way you addressed the issue of installation instructions @Nosferican ? I can't see any changes to either the readme, docs or paper on installation. Thanks.
