### 2019-08-14 #38 @Nosferican
Absolutely.
- [docs](https://nosferican.github.io/Econometrics.jl/dev/getting_started/#Installation-1) (right now in the dev, but as soon as I tag a new release it will be updated to the stable as well)
- [paper](https://github.com/JuliaCon/proceedings-papers/blob/jcon.00038/jcon.00038/10.21105.jcon.00038.pdf) I added to the conclusion how to obtain the software next to the license information.
> The package is registered in the official Julia registry (General); as such it can be installed using the Julia package manager.
- [readme](https://github.com/Nosferican/Econometrics.jl), I added a shield with the information of the latest release. I don't know if it should be more explicit that it is the release in the General registry. If you think it should have something more explicit about it being registered with General or the basic installation with the Pkg manager I would be happy to include it if deemed helpful (e.g., linking to the docs for the installation instructions). I have both Registrator.jl and TagBot.jl enabled to keep both in synch.
