### 2019-08-14 #38 @pkofod
I didnt look at the dev version of the docs, it looks fine.
