### 2019-08-16 #38 @Nosferican
Let me know if there is anything that needs addressing.
