### 2019-08-19 123d2c1
Fix polr
Fixes polr estimator
- Bug when the number of parameters /outcomes differed.
