+++ test/Project.toml
@@ -5,3 +5,4 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
 CSV = "0.5"
+RDatasets = "0.6.3"
