### 2019-08-19 88e1e26
Fix polr (#19)
Fixes polr estimator
- Bug when the number of parameters /outcomes differed.
