+++ test/runtests.jl
@@ -234,16 +234,33 @@ end
   	β, V, σ = coef(model), vcov(model), stderror(model)
   	@test β ≈ [ 0.009437926, -0.015143049, -0.103911316,
                -2.92391240,  -1.549266200, -0.298963900, 0.6698249] rtol = 1e-4
-	@test V ≈ [ 0.00000646 -0.0000007000  0.000013250 0.00031473	0.0003161358 0.000321276 0.0003270734
-               -0.00000070  0.0071633800 -0.000184930 0.00294759	0.0029647442 0.002995366 0.0029904485
-                0.00001325 -0.0001849300  0.000618730 0.00288365	0.0028548209 0.002791361 0.0027318928
-                0.00031473  0.0029475900  0.002883650 0.03684071	0.0289239290 0.026986237 0.026498765
-                0.00031614  0.0029647400  0.002854820 0.02892393	0.0293971651 0.027160433 0.0266006675
-                0.000321276 0.0029953660  0.002791361 0.02698623	0.0271604331 0.027860143 0.0270590982
-                0.0003270734 0.002990448  0.002731893 0.02649876	0.0266006675 0.027059098 0.0281002038
+	@test V ≈ [ 0.00000646  -0.0000007000  0.000013250  0.00031473  0.0003161358  0.000321276  0.0003270734
+               -0.00000070   0.0071633800 -0.000184930  0.00294759  0.0029647442  0.002995366  0.0029904485
+                0.00001325  -0.0001849300  0.000618730  0.00288365  0.0028548209  0.002791361  0.0027318928
+                0.00031473   0.0029475900  0.002883650  0.03684071  0.0289239290  0.026986237  0.026498765
+                0.00031614   0.0029647400  0.002854820  0.02892393  0.0293971651  0.027160433  0.0266006675
+                0.000321276  0.0029953660  0.002791361  0.02698623  0.0271604331  0.027860143  0.0270590982
+                0.0003270734 0.002990448   0.002731893  0.02649876  0.0266006675  0.027059098  0.0281002038
               ] rtol = 1e-3
     @test σ ≈ [0.002540789, 0.084636770, 0.024874381,
                0.191939349, 0.171456015, 0.166913580, 0.167631154] rtol = 1e-4
+	data = data[data.RecParks .≠ 5,:]
+	model = fit(EconometricModel,
+              	@formula(RecParks ~ Age + Sex + Schooling),
+              	data)
[…]
