### 2019-08-19 1fa3a0f
Fix docstring of EconometricModel
Outdated docstring.
