+++ CHANGELOG.md
@@ -1,0 +1,21 @@
+# v0.2.0
+
+## Features
+
+- Dropped DataFrames requirement in favor of the Tables.jl API ([#13](https://github.com/Nosferican/Econometrics.jl/pull/13))
+- Improved the documentation.
+
+## Bug fixes
+
+- Pretty print for the formula ([#8](https://github.com/Nosferican/Econometrics.jl/issues/8))
+- Fixed issue with the ordinal response model ([#19](https://github.com/Nosferican/Econometrics.jl/pull/19))
+- Fixed outdated docstrings
+
+## Miscellaneous
+
+- Added Windows (x86 and x64 architecture) to the continuous integration test matrix ([commit](https://github.com/Nosferican/Econometrics.jl/commit/969ea35d3f65c401bbf941c0038a3c39f88e9afd))
+- Paper and docs incorporated the feedback from the peer-review process (JuliaCon proceedings)
+
+# v0.1.0
+
+Initial public release (beta)
