### 2019-08-20 #38 @matbesancon
@Nosferican can you check the missing or invalid DOIs above?
