### 2019-08-20 aca9032
Update .travis.yml
- Add LTS, latest release, and next release to CI
- Add Coverage
