+++ .travis.yml
@@ -4,7 +4,10 @@ os:
   - osx
   - windows
 julia:
+  - 1.0
+  - 1.2
   - 1.3
+codecov: true
 notifications:
   email: false
 jobs:
