### 2019-08-20 acfb740
Clean up of ref.bib
- Clean up the ref.bib to exclude nocite
- Added note about chapter 3 of dissertation
