+++ paper/paper.tex
@@ -310,6 +310,8 @@ Community contributions and feedback are highly encouraged in order to best cont
 
 This research benefited from support from the National Science Foundation and the AEA Mentoring Program: NSF Awards 1357478 \& 1730651.
 
+I would also like to thank Hisam Sabouni for useful feedback during the development of the project and at an earlier draft of the article \cite{JBSC2019} and Jeffrey Wooldridge for guidance in the programming of the estimators.
+
 \input{bib.tex}
 
 \appendix
