+++ paper/ref.bib
@@ -1,368 +1,11 @@
-@techreport{Abadie_Athey_Imbens_Wooldridge_2017,
-	author = {Abadie, Alberto and Athey, Susan and Imbens, Guido W. and Wooldridge, Jeffrey M.},
-	title = {When Should You Adjust Standard Errors for Clustering?},
-	institution = {National Bureau of Economic Research},
-	type = {Working Paper},
-	series = {Working Paper Series},
-	number = {24003},
-	year = {2017},
-	doi = {10.3386/w24003}
-}
-@article{Arellano_1987,
-	author = {Arellano, Manuel},
-	title = {Computing Robust Standard Errors for Within-groups Estimators},
-	journal =  {Oxford Bulletin of Economics and Statistics},
-	volume = {49},
-	pages = {431--434},
-	number = {4},
-	year = {1987},
-	doi = {10.1111/j.1468-0084.1987.mp49004006.x}
-}
-@article{Athey_Imbens_2016,
-	author = {Athey, Susan and Imbens, Guido W.},
-	title = {The Econometrics of Randomized Experiments},
-	journal =  {ArXiv e-prints},
-	archivePrefix = {arXiv},
-	eprint = {1607.00698},
-	primaryClass = {stat.ME},
-	keywords = {Statistics - Methodology, Economics - Econometrics, 62K99, G.3},
-	year = {2016},
-	url = {https://arxiv.org/abs/1607.00698}
-	}
-@book{Baltagi_2013,
-	location = {Chichester, West Sussex},
-	edition = {Fifth Edition},
-	title = {Econometric analysis of panel data},
-	isbn = {978-1-118-67232-7},
-	publisher = {John Wiley \& Sons, Inc},
-	author = {Baltagi, Badi H.},
-	date = {2013}
-}
-@article{Blinder_1973,
-	title = {Wage Discrimination: Reduced Form and Structural Estimates},
-	volume = {8},
-	issn = {0022166X},
-	doi = {10.2307/144855},
-	shorttitle = {Wage Discrimination},
-	pages = {436},
-	number = {4},
-	journal = {The Journal of Human Resources},
-	author = {Blinder, Alan S.},
-	date = {1973}
-}
-@article{Cameron_Gelbach_Miller_2011,
-    author = {Cameron, Adrian C. and Gelbach, Jonah B. and Miller, Douglas L.},
-    journal =  {Journal of Business \& Economic Statistics},
-    number = {2},
-    pages = {238--249},
-    publisher = {American Statistical Association},
-    title = {Robust Inference With Multiway Clustering},
-    volume = {29},
-    year = {2011},
-    doi = {10.2307/25800796}
-    }
-@article{Cameron_Miller_2015,
-	author = {Cameron, A. Colin and Miller, Douglas L.},
-	title = {A Practitioner's Guide to Cluster-Robust Inference},
-	journal = {Journal of Human Resources},
-	volume = {50},
-	pages = {317--372},
-	number = {2},
-	year = {2015},
-	doi = {10.3368/jhr.50.2.317}
-}
-@article{Card_Krueger_1992,
-	author = {Card, David and Krueger, Alan B.},
-	title = {School Quality and Black-White Relative Earnings: A Direct Assessment*},
-	journal = {The Quarterly Journal of Economics},
-	volume = {107},
-	number = {1},
-	pages = {151-200},
-	year = {1992},
-	doi = {10.2307/2118326}
-}
-@article{Dettori_Norvell_Skelly_Chapman_2011,
-	author = {Dettori, Joseph and Norvell, Daniel and Skelly, Andrea and Chapman, Jens},
-	title = {Heterogeneity of treatment effects: from ``How to treat'' to ``Whom to treat''},
-	volume = {2},
-	issn = {1663-7976, 1869-4136},
[…]
