### 2019-08-20 df47016
Support for LTS version
Add compat for ismissing/isnothing
