+++ Project.toml
@@ -25,4 +25,4 @@ Optim = "0.18"
 Parameters = "0.10"
 StatsBase = "0.31"
 StatsModels = "0.6"
-julia = "1.1"
+julia = "1"
