+++ test/Project.toml
@@ -1,8 +1,0 @@
-[deps]
-CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
-RDatasets = "ce6b1742-4840-55fa-b093-852dadbb1d8b"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[compat]
-CSV = "0.5"
-RDatasets = "0.6.3"
