### 2019-08-20 2c7253a
Update .travis.yml
Drop Julia 1.3
