+++ .travis.yml
@@ -6,7 +6,6 @@ os:
 julia:
   - 1.0
   - 1.2
-  - 1.3
 codecov: true
 notifications:
   email: false
@@ -14,10 +13,13 @@ jobs:
   include:
     - os: windows
       arch: x86
-      julia: 1.3
+      julia: 1.0
+    - os: windows
+      arch: x86
+      julia: 1.2
     - stage: "Documentation"
       os: linux
-      julia: 1.3
+      julia: 1.2
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
                                                Pkg.instantiate()'
