### 2019-08-20 3a5d2d0
Julia LTS release support (#20)
* Support for LTS release
