+++ Project.toml
@@ -20,9 +20,19 @@ StatsModels = "3eaba693-59b7-5ba5-a881-562e759f1c8d"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
+CSV = "0.5"
 FillArrays = "0.6"
 Optim = "0.18"
 Parameters = "0.10"
+RDatasets = "0.6.3"
 StatsBase = "0.31"
 StatsModels = "0.6"
 julia = "1"
+
+[extras]
+CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
+RDatasets = "ce6b1742-4840-55fa-b093-852dadbb1d8b"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["CSV", "RDatasets", "Test"]
