### 2019-08-21 #38 @matbesancon
@Nosferican some remarks:
Page 2:
footnote 1: there is no A and b here, so it may be hard to understand what you mean. Something like: " `\` is a linear solver operator, such that `x = A\b => Ax = b`
Equation 3: I assume `.*` is a broadcasted multiplication operator? It should be specified.

Page 3:
"solving these problemS" (typo)
Equation 11: maybe without using a double letter for a single variable? It can be confusing
Section 2.5 title: First-Difference Estimator (not just First-Difference)

Reference [2]: use double bracket `{{}}` to maintain the capitalized MixedModels.jl
