### 2019-08-21 #38 @matbesancon
Best practices: you can add a footnote link to the documentation
