### 2019-08-21 #38 @matbesancon
Page 4:
> The Julia ecosystem follows this canonical pipeline.

The sentence is confusing, the Julia statistics and data ecosystems enable / provide features for the whole pipeline?
