### 2019-08-21 #38 @matbesancon
Don't forget to properly cite the packages you mention.
This is more of a personal opinion, so you may take it with a grain of salt, but I find the "Design Decisions" section to be a bit informal. I would rephrase the "Should software be dummy-proof?" question for instance
