### 2019-08-22 #38 @Nosferican
Thanks for the feedback. I think I have gone over all the issues. For packages that do not provide any suggested citation such as no Zedeno/article (e.g., StatsModels.jl / DataFrames.jl), any suggestion on how to cite them?
