### 2019-08-23 #38 @matbesancon
What I usually do is something like:
```
@misc{kerneldensity,
    author = {{JuliaStats}},
    title = {{KernelDensity.jl}},
    url = {https://github.com/JuliaStats/KernelDensity.jl},
    version = {0.4.1},
    year = {2018},
    month= {02}
}
```

Feel free to replace the authors with the maintainers' names
