### 2019-08-26 #38 @matbesancon
Also, I think you haven't put the package citations
