### 2019-08-26 f6df0f1
Added Survival.jl manual reference.
