+++ paper/paper.tex
@@ -228,7 +228,7 @@ StatsModels.jl \cite{StatsModels.jl} is a package that provides the means to go 
 \subsection{Regression Analysis}
 
 The regression analysis ecosystem in Julia has GLM.jl \cite{GLM.jl} as its flagship. GLM.jl provides the typical functionality for fitting generalized linear models through Fisher scoring. This includes linear models, Poisson/Negative Binomial, Logit/Probit, and other non-canonical link models. CovarianceMatrices.jl \cite{CovarianceMatrices.jl} provides various variance covariance estimators for GLM.jl models à la R's
-sandwich package. LinearMixedModels.jl \cite{LinearMixedModels} extends GLM.jl for mixed-effects models. FixedEffectModels.jl provides fast estimation of linear models with instrumental variables and high dimensional categorical variables à la reghdfe. Survival.jl provides a series of estimators for duration models. Two major gaps in the ecosystem include estimating nominal and ordinal response models (i.e., discrete choice) with more than two alternatives and support for longitudinal estimators.
+sandwich package. LinearMixedModels.jl \cite{LinearMixedModels} extends GLM.jl for mixed-effects models. FixedEffectModels.jl provides fast estimation of linear models with instrumental variables and high dimensional categorical variables à la reghdfe. Survival.jl \cite{Survival.jl} provides a series of estimators for duration models. Two major gaps in the ecosystem include estimating nominal and ordinal response models (i.e., discrete choice) with more than two alternatives and support for longitudinal estimators.
 
 \section{Econometrics.jl}
 
