+++ paper/ref.bib
@@ -303,7 +303,7 @@
   url          = {https://github.com/JuliaStats/GLM.jl}
 }
 @misc{StatsModels.jl,
-  author       = {Kleinschmidt, Dave and
+  author       = {Kleinschmidt, Dave F. and
                   White, John Myles and
                   Bouchet-Valat, Milan and
                   Bates, Douglas and
@@ -430,4 +430,12 @@
   title        = {{FixedEffectModels.jl}},
   year         = 2015,
   url          = {https://github.com/matthieugomez/FixedEffectModels.jl}
+}
+@misc{Survival.jl,
+  author       = {Arslan, Alex and
+                  Kleinschmidt, Dave F. and
+                  Vertechi, Pietro},
+  title        = {{Survival.jl}},
+  year         = 2017,
+  url          = {https://github.com/JuliaStats/Survival.jl}
 }
