### 2019-08-27 #38 @matbesancon
there is a weird effect of two columns almost merging in the middle. @Nosferican do you get the same effect locally?
