### 2019-08-27 c87e391
Updates from Overleaf
