### 2019-08-27 #38 @matbesancon
ok @Nosferican this looks good to me, you can do a last check, the last things:
I would advise releasing a new version, then putting it on Zenodo to get a DOI, the DOI will be attached to the paper

Once this is done, I will set the version, DOI and figure out how to publish
