+++ paper/paper.tex
@@ -228,7 +228,7 @@ StatsModels.jl \cite{StatsModels.jl} is a package that provides the means to go 
 \subsection{Regression Analysis}
 
 The regression analysis ecosystem in Julia has GLM.jl \cite{GLM.jl} as its flagship. GLM.jl provides the typical functionality for fitting generalized linear models through Fisher scoring. This includes linear models, Poisson/Negative Binomial, Logit/Probit, and other non-canonical link models. CovarianceMatrices.jl \cite{CovarianceMatrices.jl} provides various variance covariance estimators for GLM.jl models à la R's
-sandwich package. LinearMixedModels.jl \cite{LinearMixedModels} extends GLM.jl for mixed-effects models. FixedEffectModels.jl provides fast estimation of linear models with instrumental variables and high dimensional categorical variables à la reghdfe. Survival.jl \cite{Survival.jl} provides a series of estimators for duration models. Two major gaps in the ecosystem include estimating nominal and ordinal response models (i.e., discrete choice) with more than two alternatives and support for longitudinal estimators.
+sandwich package. MixedModels.jl \cite{MixedModels.jl} extends GLM.jl for mixed-effects models. FixedEffectModels.jl provides fast estimation of linear models with instrumental variables and high dimensional categorical variables à la reghdfe. Survival.jl \cite{Survival.jl} provides a series of estimators for duration models. Two major gaps in the ecosystem include estimating nominal and ordinal response models (i.e., discrete choice) with more than two alternatives and support for longitudinal estimators.
 
 \section{Econometrics.jl}
 
@@ -240,7 +240,7 @@ This section will showcase some examples of using the package for various estima
 
 For linear models, the examples use a crime dataset \cite{Crime}. The data set is a balanced longitudinal data set with 90 counties in North Carolina from 1981 to 1987. The outcome variable is the crime rate and the explanatory variables include the probability of conviction, average sentence, and probability of prison sentence. Estimating the pooling estimator or the between estimator can be accomplished as in figure \vref{estimator_pooling} and \vref{estimator_between}. Table \vref{poberes} shows the estimated 95\% confidence intervals using Econometrics.jl, Stata, and R's plm package.
 
-The syntax to estimate a model follows the StatsBase API which provides two methods for building and estimating models (i.e., \textit{fit} and \textit{fit!}). The \textit{fit} approach takes (1) an estimator, (2) a model formula, (3) data, and additional arguments depending on the model. \textit{EconometricModel} is the generic estimator which dispatches to continuous, nominal or ordinal response models automatically based on the type of the response. Other estimators are available and the following figures illustrate some of the use cases.
+The syntax to estimate a model follows the StatsBase API which provides a method for building and estimating models (i.e., \textit{fit}). The \textit{fit} approach takes (1) an estimator, (2) a model formula, (3) data, and additional arguments depending on the model. \textit{EconometricModel} is the generic estimator which dispatches to continuous, nominal or ordinal response models automatically based on the type of the response. Other estimators are available and the following figures illustrate some of the use cases.
 
 The fixed effects model or within estimator can be estimated as in figure \vref{withinmodel} which a two-ways fixed effects model (i.e., fixed effects for panel and time dimensions). Table \vref{withinres} shows the estimated 95\% confidence intervals using Econometrics.jl, Stata, and R's plm package.
 
