### 2019-08-28 41ddabe
v0.2.2
Added support for Julia v1.0
