+++ Project.toml
@@ -3,7 +3,7 @@ uuid = "4d6a76a9-bfbc-5492-8924-cf6ed7875f06"
 license = "ISC"
 description = "Econometrics in Julia"
 maintainer = "José Bayoán Santiago Calderón"
-version = "0.2.0"
+version = "0.2.2"
 
 [deps]
 CategoricalArrays = "324d7699-5711-5eae-9e2f-1d82baa6b597"
