### 2019-08-28 #38 @matbesancon
that's bizarre. Are you using this `csquotes.sty`?
