### 2019-08-28 #38 @matbesancon
could you add it to the dependencies, it might not be included in the server's build
