### 2019-08-28 #38 @matbesancon
In the `paper` folder, simply add the corresponding `sty` file which should be easy to find online
