### 2019-08-28 #38 @matbesancon
I think csquotes.def and *.cfg are also required, sorry!
