+++ paper/figures/polr.tex
@@ -1,5 +1,5 @@
 \begin{table}[hbpt]
-    \caption{Parallel Ordinal Logistic Regression}
+    \caption{Proportional Odds Logistic Regression}
     \label{ologitres}
     \begin{threeparttable}
     \begin{tabularx}{\textwidth}{Y c c c c c c}
