### 2019-10-09 f71a8a7
Update dependencies and make them more flexible
