### 2019-10-09 47c5f5c
Apply suggestions from code review
