+++ Project.toml
@@ -24,7 +24,7 @@ CSV = "0.5"
 FillArrays = "0.6, 0.7"
 Optim = "0.18, 0.19"
 Parameters = "0.10"
-RDatasets = "0.6.3, 0.6.4"
+RDatasets = "0.6.3"
 StatsBase = "0.31, 0.32"
 StatsModels = "0.6"
 julia = "1"
