### 2019-10-09 9a1e12f
Update dependencies
* Update dependencies
Will fix the docs file issue and tag the patch version.
