+++ Project.toml
@@ -3,7 +3,7 @@ uuid = "4d6a76a9-bfbc-5492-8924-cf6ed7875f06"
 license = "ISC"
 description = "Econometrics in Julia"
 maintainer = "José Bayoán Santiago Calderón"
-version = "0.2.2"
+version = "0.2.3"
 
 [deps]
 CategoricalArrays = "324d7699-5711-5eae-9e2f-1d82baa6b597"
@@ -21,11 +21,11 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
 CSV = "0.5"
-FillArrays = "0.6"
-Optim = "0.18"
+FillArrays = "0.6, 0.7"
+Optim = "0.18, 0.19"
 Parameters = "0.10"
 RDatasets = "0.6.3"
-StatsBase = "0.31"
+StatsBase = "0.31, 0.32"
 StatsModels = "0.6"
 julia = "1"
 
