### 2019-10-10 a33a427
Clean the docs
- Having an issue where VCE source has the wrong link and is not being shown in the correct block.
- confint shows multiple times separate docstring instead of the merged one.
- In another issue I will check the poly2 example.
