### 2019-10-10 2765919
Clean docs (#25)
- Using docs since autodocs is kinda
- I also fixed bug with special terms
