+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"public_api/#Public-API-1","page":"Public API","title":"Public API","text":"","category":"section"},{"location":"public_api/#","page":"Public API","title":"Public API","text":"absorb\nEconometricModel\nBetweenEstimator\nRandomEffectsEstimator\nEconometrics.VCE","category":"page"},{"location":"public_api/#Econometrics.absorb","page":"Public API","title":"Econometrics.absorb","text":"absorb\n\nFunction for constructing the FunctionTerm{typeof(absorb)} used in decompose.\n\n\n\n\n\n","category":"function"},{"location":"public_api/#Econometrics.EconometricModel","page":"Public API","title":"Econometrics.EconometricModel","text":"EconometricModel(estimator::Type{<:Union{EconometricModel,ModelEstimator}},\n                 f::FormulaTerm,\n                 data;\n                 contrasts::Dict{Symbol} = Dict{Symbol,Union{<:AbstractContrasts,<:AbstractTerm}}(),\n                 wts::Union{Nothing,Symbol} = nothing,\n                 panel::Union{Nothing,Symbol} = nothing,\n                 time::Union{Nothing,Symbol} = nothing,\n                 vce::VCE = OIM)\n\nFormula has syntax:\n\n@formula(response ~ exogenous + (endogenous ~ instruments) + absorb(highdimscontrols))\n\nData must implement the Tables.jl API and use CategoricalArrays (CategoricalVector)\n\nWeights are taken as StatsBase.FrequencyWeights\n\nPanel and time indicators are used for longitudinal estimators\n\nExamples\n\nmodel = fit(EconometricModel, formula, data, kwargs...)\nmodel = fit(BetweenEstimator, formula, data, panel = :panel, kwargs...)\nmodel = fit(RandomEffectsEstimator, formula, data, panel = :panel, time = :time, kwargs...)\n\n\n\n\n\n","category":"type"},{"location":"public_api/#Econometrics.BetweenEstimator","page":"Public API","title":"Econometrics.BetweenEstimator","text":"BetweenEstimator(effect::Symbol,\n                 groups::Vector{Vector{Int}}) <: LinearModelEstimators\n\nContinuous response estimator collapsing a dimension in a longitudinal setting.\n\n\n\n\n\n","category":"type"},{"location":"public_api/#Econometrics.RandomEffectsEstimator","page":"Public API","title":"Econometrics.RandomEffectsEstimator","text":"RandomEffectsEstimator(pid::Tuple{Symbol,Vector{Vector{Int}}},\n                       tid::Tuple{Symbol,Vector{Vector{Int}}},\n                       idiosyncratic::Float64,\n                       individual::Float64,\n                       θ::Vector{Float64}) <: LinearModelEstimators\n\nSwamy-Arora estimator.\n\n\n\n\n\n","category":"type"},{"location":"public_api/#Econometrics.VCE","page":"Public API","title":"Econometrics.VCE","text":"VCE\n\nVariance-covariance estimators.\n\nObserved Information Matrix (OIM)\nHeteroscedasticity Consistent: HC0, HC1, HC2, HC3, HC4\n\n\n\n\n\n","category":"type"},{"location":"getting_started/#Getting-Started-1","page":"Getting Started","title":"Getting Started","text":"","category":"section"},{"location":"getting_started/#Installation-1","page":"Getting Started","title":"Installation","text":"","category":"section"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"Econometrics.jl is registered with the Julia's General registry.","category":"page"},{"location":"getting_started/#","page":"Getting Started","title":"Getting Started","text":"For installing the package, users 
[…]
