### 2019-11-06 c571ce3
Update Parameters Compat entry (#28)
Parameters 0.10, 0.11, and 0.12.
