+++ Project.toml
@@ -23,7 +23,7 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 CSV = "0.5"
 FillArrays = "0.6, 0.7"
 Optim = "0.18, 0.19"
-Parameters = "0.10"
+Parameters = "0.10, 0.11, 0.12"
 RDatasets = "0.6.3"
 StatsBase = "0.31, 0.32"
 StatsModels = "0.6"
