### 2019-11-06 2b55857
Patch release for compat
@JuliaRegistrator register

Release notes:

Updated Parameters.jl compat entry for including 0.11 and 0.12.
