### 2019-11-06 d1c8e47
Added compat entries.
