+++ Project.toml
@@ -20,13 +20,18 @@ StatsModels = "3eaba693-59b7-5ba5-a881-562e759f1c8d"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
+CategoricalArrays = "0.6, 0.7"
 CSV = "0.5"
+Distributions = "0.21"
 FillArrays = "0.6, 0.7"
+ForwardDiff = "0.10"
 Optim = "0.18, 0.19"
 Parameters = "0.10, 0.11, 0.12"
 RDatasets = "0.6.3"
 StatsBase = "0.31, 0.32"
+StatsFuns = "0.7, 0.8"
 StatsModels = "0.6"
+Tables = "0.2"
 julia = "1"
 
 [extras]
