### 2020-03-24 501891c
Expand compat entries
