+++ Project.toml
@@ -22,16 +22,16 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 [compat]
 CategoricalArrays = "0.6, 0.7"
 CSV = "0.5"
-Distributions = "0.21"
-FillArrays = "0.6, 0.7"
+Distributions = "0.21, 0.22, 0.23"
+FillArrays = "0.6, 0.7, 0.8"
 ForwardDiff = "0.10"
-Optim = "0.18, 0.19"
+Optim = "0.18, 0.19, 0.20"
 Parameters = "0.10, 0.11, 0.12"
 RDatasets = "0.6.3"
 StatsBase = "0.31, 0.32"
-StatsFuns = "0.7, 0.8"
+StatsFuns = "0.7, 0.8, 0.9"
 StatsModels = "0.6"
-Tables = "0.2"
+Tables = "0.2, 1"
 julia = "1"
 
 [extras]
