### 2020-03-24 3d3ec1e
Add GitHub Actions CI
