### 2020-03-24 b7e20d8
Delete .travis.yml
Added the GitHub Actions C.I.
