+++ .travis.yml
@@ -1,27 +1,0 @@
-language: julia
-os:
-  - linux
-  - osx
-  - windows
-julia:
-  - 1.0
-  - 1.2
-codecov: true
-notifications:
-  email: false
-jobs:
-  include:
-    - os: windows
-      arch: x86
-      julia: 1.0
-    - os: windows
-      arch: x86
-      julia: 1.2
-    - stage: "Documentation"
-      os: linux
-      julia: 1.2
-      script:
-        - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
-                                               Pkg.instantiate()'
-        - julia --project=docs/ docs/make.jl
-      after_success: skip
