+++ src/formula.jl
@@ -28,64 +28,85 @@ Decomposes the arguments passed to a fit(::EconometricModel) into its components
 - Z::Matrix{Float64}
 - wts::FrequencyWeights
 """
-function decompose(f::FormulaTerm,
-                   data,
-                   contrasts::Dict{Symbol},
-                   wts::Union{Nothing,Symbol},
-                   panel::Union{Nothing,Symbol},
-                   time::Union{Nothing,Symbol},
-                   estimator::Type{<:Union{EconometricsModel,ModelEstimator}},
-                   vce::VCE)
+function decompose(
+    f::FormulaTerm,
+    data,
+    contrasts::Dict{Symbol},
+    wts::Union{Nothing,Symbol},
+    panel::Union{Nothing,Symbol},
+    time::Union{Nothing,Symbol},
+    estimator::Type{<:Union{EconometricsModel,ModelEstimator}},
+    vce::VCE,
+)
     rhs = isa(f.rhs, Tuple) ? collect(f.rhs) : [f.rhs]
     absorbed = findall(t -> isa(t, FunctionTerm{typeof(absorb)}), rhs)
     # Conditions based on panel and temporal indices and absorbed features
     if isa(estimator, Type{<:BetweenEstimator}) && isnothing(panel)
         throw(ArgumentError("The between estimator requires the panel identifier."))
-    elseif isa(estimator, Type{<:RandomEffectsEstimator}) && (isnothing(panel) || isnothing(time))
+    elseif isa(estimator, Type{<:RandomEffectsEstimator}) &&
+           (isnothing(panel) || isnothing(time))
         throw(ArgumentError("The random effects estimator requires the panel and temporal identifiers."))
-    elseif (isa(estimator, Type{<:BetweenEstimator}) ||
-            isa(estimator, Type{<:RandomEffectsEstimator})) && !isempty(absorbed)
+    elseif (
+        isa(estimator, Type{<:BetweenEstimator}) ||
+        isa(estimator, Type{<:RandomEffectsEstimator})
+    ) && !isempty(absorbed)
         throw(ArgumentError("Absorbing features is only implemented for least squares."))
     end
-    pns = convert(Vector{Symbol}, filter!(!isnothing, union(termvars(f), [panel, time, wts])))
-    data = select(data, pns...)
+    pns =
+        convert(Vector{Symbol}, filter!(!isnothing, union(termvars(f), [panel, time, wts])))
+    data = select(data, pns...) |> columntable
     if isa(wts, Symbol)
-		wts = ifelse.(getproperty(data, wts) .≤ 0, missing, getproperty(data, wts))
+        wts = ifelse.(getproperty(data, wts) .≤ 0, missing, getproperty(data, wts))
     end
-	categorical_variables = Tables.schema(data) |>
-		(x -> [name for (name, type) ∈ zip(x.names, x.types) if
-			isa(type, Type{<:Union{AbstractString, CategoricalValue}})])
-	data_categorical =
-		zip(categorical_variables,
-			(levels(col), isordered(col)) for
-		    col ∈ eachcolumn(select(data, categorical_variables...))) |>
-		Dict
-	data = materializer(data)(row for row ∈ rows(data) if
-        all(pn -> !ismissing(getproperty(row, pn)), propertynames(row)))
-    for cn ∈ categorical_variables
-		col = getproperty(data, cn)
-		levels!(col, intersect(data_categorical[cn][1],
-							   unique(col)))
-		ordered!(col, data_categorical[cn][2])
-	end
+    categorical_variables =
+        Tables.schema(data) |> (
+            x -> [
+                name
+                for
+                (name, type) ∈ zip(x.names, x.types) if
[…]
