+++ Project.toml
@@ -17,21 +17,19 @@ Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 StatsModels = "3eaba693-59b7-5ba5-a881-562e759f1c8d"
-Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
+TableOperations = "ab02a1b2-a7df-11e8-156e-fb1833f50b87"
 
 [compat]
 CategoricalArrays = "0.6, 0.7"
-CSV = "0.5"
 Distributions = "0.21, 0.22, 0.23"
 FillArrays = "0.6, 0.7, 0.8"
 ForwardDiff = "0.10"
 Optim = "0.18, 0.19, 0.20"
 Parameters = "0.10, 0.11, 0.12"
-RDatasets = "0.6.3"
 StatsBase = "0.31, 0.32"
 StatsFuns = "0.7, 0.8, 0.9"
 StatsModels = "0.6"
-Tables = "0.2, 1"
+TableOperations = "0.2"
 julia = "1"
 
 [extras]

[…]
