+++ src/transformations.jl
@@ -5,26 +5,29 @@
 
 This function performs the within transformation given a model matrix and fixed effects using the method of alternating projections.
 """
-@views within(obj::AbstractMatrix{<:Real},
-              D::AbstractVector{<:AbstractVector{<:AbstractVector{<:Integer}}},
-              wts::AbstractVector) =
-    isempty(D) ? obj : mapslices(col -> within(col, D, wts), obj, dims = 1)
-@views function within(obj::AbstractVector{<:Real},
-                       D::AbstractVector{<:AbstractVector{<:AbstractVector{<:Integer}}},
-                       wts::AbstractVector = Ones(length(obj)))
+@views within(
+    obj::AbstractMatrix{<:Real},
+    D::AbstractVector{<:AbstractVector{<:AbstractVector{<:Integer}}},
+    wts::AbstractVector,
+) = isempty(D) ? obj : mapslices(col -> within(col, D, wts), obj, dims = 1)
+@views function within(
+    obj::AbstractVector{<:Real},
+    D::AbstractVector{<:AbstractVector{<:AbstractVector{<:Integer}}},
+    wts::AbstractVector = Ones(length(obj)),
+)
     isempty(D) && return obj
     current = copy(obj)
     output = copy(obj)
     er = Inf
     μ = mean(obj, wts)
     while er > 1e-8
-        for dimension ∈ D
+        for dimension in D
             current = copy(output)
-            for group ∈ dimension
+            for group in dimension
                 output[group] .-= mean(current[group], FrequencyWeights(wts[group]))
             end
         end
-        er = √sum((x - y)^2 for (x, y) ∈ zip(output, current))
+        er = √sum((x - y)^2 for (x, y) in zip(output, current))
     end
     output .+= μ
 end
@@ -35,18 +38,21 @@ end
 
 This function performs the partial within transformation given a model matrix subgroups and subgroup specific error components.
 """
-@views function partialwithin(obj::AbstractVecOrMat{<:Real},
-                              panel::AbstractVector{<:AbstractVector{<:Integer}},
-                              θ::AbstractVector{<:Real},
-                              wts::AbstractVector)
+@views function partialwithin(
+    obj::AbstractVecOrMat{<:Real},
+    panel::AbstractVector{<:AbstractVector{<:Integer}},
+    θ::AbstractVector{<:Real},
+    wts::AbstractVector,
+)
     output = copy(obj)
     if !isempty(wts)
-        for (panel, θ) ∈ zip(panel, θ)
-            output[panel,:] .-= θ * mean(obj[panel, :], FrequencyWeights(wts[panel]), dims = 1)
+        for (panel, θ) in zip(panel, θ)
+            output[panel, :] .-=
+                θ * mean(obj[panel, :], FrequencyWeights(wts[panel]), dims = 1)
         end
     else
-        for (panel, θ) ∈ zip(panel, θ)
-            output[panel,:] .-= θ * mean(obj[panel, :], dims = 1)
+        for (panel, θ) in zip(panel, θ)
+            output[panel, :] .-= θ * mean(obj[panel, :], dims = 1)
         end
     end
     output
@@ -73,22 +79,30 @@ function transform end
 transform(estimator::LinearModelEstimators, wts::FrequencyWeights) = wts
 transform(estimator::BetweenEstimator, wts::FrequencyWeights) =
     FrequencyWeights(ones(length(estimator.groups)))
-transform(estimator::ContinuousResponse,
-          obj::AbstractVecOrMat{<:Number},
[…]
