+++ Project.toml
@@ -26,7 +26,7 @@ FillArrays = "0.6, 0.7, 0.8"
 ForwardDiff = "0.10"
 Optim = "0.18, 0.19, 0.20"
 Parameters = "0.10, 0.11, 0.12"
-StatsBase = "0.31, 0.32"
+StatsBase = "0.31, 0.32, 0.33"
 StatsFuns = "0.7, 0.8, 0.9"
 StatsModels = "0.6"
 TableOperations = "0.2"
