### 2020-03-27 b5a8c9e
Bump compat StatsBase
