### 2020-03-27 8d3a206
CompatHelper: bump compat for "StatsBase" to "0.33" (#38)
* CompatHelper: bump compat for "StatsBase" to "0.33"

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: José Bayoán Santiago Calderón <naoyabpr@gmail.com>
