+++ Project.toml
@@ -3,7 +3,7 @@ uuid = "4d6a76a9-bfbc-5492-8924-cf6ed7875f06"
 license = "ISC"
 description = "Econometrics in Julia"
 maintainer = "José Bayoán Santiago Calderón"
-version = "0.2.5"
+version = "0.2.6"
 
 [deps]
 CategoricalArrays = "324d7699-5711-5eae-9e2f-1d82baa6b597"
@@ -26,7 +26,7 @@ FillArrays = "0.6, 0.7, 0.8"
 ForwardDiff = "0.10"
 Optim = "0.18, 0.19, 0.20"
 Parameters = "0.10, 0.11, 0.12"
-StatsBase = "0.31, 0.32"
+StatsBase = "0.31, 0.32, 0.33"
 StatsFuns = "0.7, 0.8, 0.9"
 StatsModels = "0.6"
 TableOperations = "0.2"
