+++ Project.toml
@@ -24,7 +24,7 @@ CategoricalArrays = "0.6, 0.7"
 Distributions = "0.21, 0.22, 0.23"
 FillArrays = "0.6, 0.7, 0.8"
 ForwardDiff = "0.10"
-Optim = "0.18, 0.19, 0.20"
+Optim = "0.18, 0.19, 0.20, 0.21"
 Parameters = "0.10, 0.11, 0.12"
 StatsBase = "0.31, 0.32, 0.33"
 StatsFuns = "0.7, 0.8, 0.9"
