### 2020-06-13 445d667
bumping versions of CategoricalArrays,Optim
