### 2020-06-14 95a5393
CompatHelper: bump compat for "CategoricalArrays" to "0.8"
