+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#Interacting-with-a-model-1","page":"API","title":"Interacting with a model","text":"","category":"section"},{"location":"api/#","page":"API","title":"API","text":"The API allows for:","category":"page"},{"location":"api/#","page":"API","title":"API","text":"build/estimate models\nquery models for statistics (e.g., goodness-of-fit statistics), components (parameter estimates)\nexport model estimates","category":"page"},{"location":"api/#Core-Functionality-1","page":"API","title":"Core Functionality","text":"","category":"section"},{"location":"api/#","page":"API","title":"API","text":"Econometrics\nEconometricModel","category":"page"},{"location":"api/#Econometrics","page":"API","title":"Econometrics","text":"Econometrics\n\nEconometrics in Julia.\n\n\n\n\n\n","category":"module"},{"location":"api/#Econometrics.EconometricModel","page":"API","title":"Econometrics.EconometricModel","text":"EconometricModel(estimator::Type{<:Union{EconometricModel,ModelEstimator}},\n                 f::FormulaTerm,\n                 data;\n                 contrasts::Dict{Symbol} = Dict{Symbol,Union{<:AbstractContrasts,<:AbstractTerm}}(),\n                 wts::Union{Nothing,Symbol} = nothing,\n                 panel::Union{Nothing,Symbol} = nothing,\n                 time::Union{Nothing,Symbol} = nothing,\n                 vce::VCE = OIM)\n\nFormula has syntax:\n\n@formula(response ~ exogenous + (endogenous ~ instruments) + absorb(highdimscontrols))\n\nData must implement the Tables.jl API and use CategoricalArrays (CategoricalVector)\n\nWeights are taken as StatsBase.FrequencyWeights\n\nPanel and time indicators are used for longitudinal estimators\n\nExamples\n\nmodel = fit(EconometricModel, formula, data, kwargs...)\nmodel = fit(BetweenEstimator, formula, data, panel = :panel, kwargs...)\nmodel = fit(RandomEffectsEstimator, formula, data, panel = :panel, time = :time, kwargs...)\n\n\n\n\n\n","category":"type"},{"location":"api/#Statistical/Regression-Model-Abstraction-1","page":"API","title":"Statistical/Regression Model Abstraction","text":"","category":"section"},{"location":"api/#","page":"API","title":"API","text":"aic\naicc\nbic\nr2\nadjr2\nmss\nrss\ndeviance\nnulldeviance\nloglikelihood\nnullloglikelihood\ncoef\ndof\ndof_residual\ncoefnames(::EconometricModel)\ncoeftable(::EconometricModel)\nislinear\ninformationmatrix\nvcov(::EconometricModel)\nstderror(::EconometricModel, ::Econometrics.VCE)\nconfint(::EconometricModel)\nhasintercept\nisfitted\nfit(::EconometricModel)\nfit!(::EconometricModel)\nresponse\nmeanresponse\nfitted\npredict\nmodelmatrix\nresiduals\nleverage\nnobs\nweights(::EconometricModel)","category":"page"},{"location":"api/#StatsBase.aic","page":"API","title":"StatsBase.aic","text":"aic(obj::StatisticalModel)\n\nAkaike's Information Criterion, defined as -2 log L + 2k, with L the likelihood of the model, and k its number of consumed degrees of freedom (as returned by dof).\n\n\n\n\n\n","category":"function"},{"location":"api/#StatsBase.aicc","page":"API","title":"StatsBase.aicc","text":"aicc(obj::StatisticalModel)\n\nCorrected Akaike's Information Criterion for small sample sizes (Hurvich and Tsai 1989), defined 
[…]
