### 2020-06-29 #38 @Nosferican
Any word on when issue 1 will be out?
