### 2020-07-26 #38 @arfon
@Nosferican - could you please merge this PR? https://github.com/Nosferican/Econometrics.jl/pull/49
