+++ Project.toml
@@ -22,7 +22,7 @@ TableOperations = "ab02a1b2-a7df-11e8-156e-fb1833f50b87"
 [compat]
 CategoricalArrays = "0.6, 0.7, 0.8"
 Distributions = "0.21, 0.22, 0.23"
-FillArrays = "0.6, 0.7, 0.8"
+FillArrays = "0.6, 0.7, 0.8, 0.9"
 ForwardDiff = "0.10"
 Optim = "0.18, 0.19, 0.20, 0.21"
 Parameters = "0.10, 0.11, 0.12"
